Disable 'uvm_report_mock' tests for UVM 1.2
'uvm_report_mock' must have gotten busted during the time the tests were
broken. It's out of the scope of this activity to fix it, as we're only
focused on the testing setup.

Xcelium 20.09 (used for testing) uses UVM 1.2 by default, hence we have
to disable the 'test_mock_uvm_report' test as well.