Fix 'sim_8' test
It's intention was probably to check the handling of file lists.

It's probably not important that the unit test get created every time,
but this is what the original test did. If this is the case, we don't
need to store the unit test in the repo.

Since we compile 'another_filelist_module.sv' through 'a_filelist.f', we
don't need to specify it in 'svunit.f'.