Add workaround for QuestaSim issue w.r.t. local constructor
QuestaSim complains if we use the local constructor to implement eager
initialization of the singleton instance. It seems to support the lazy
initialization variant of the pattern (tried on EDAPlayground), so that is what
we'll use.

Lazy initialization is overkill for us, since we anyway get the singleton
instance as part of the static initialization of the package variable.