Add test for partial wildcard at beginning of string
This test fails, because `inside` for string comparison doesn't do what I
expect it to do. It's not immediately clear to me what `inside` does when used
for string comparison. It probably does some casting to `byte`.

Interestingly, QuestaSim and VCS don't return `1` for `"*" inside { "foo*" }`,
but Xcelium does. It's not that Xcelium searches for the `"*"` character though,
because if `"*"` is at the beginning of the string Xcelium will also return `0`
for the operation.