Make sure test code is elaborated using scripting
Making the test package a "top" instructs the simulator to elaborate it and perform any code related to static initialization. The order of tops isn't important (though the first one defines the snapshot image - here it's nice to see "svunit" used as a name). Not that this might only work in Xcelium.

The scripting could comfortably handle the task of making test packages tops. Come to think of it, scripting could also create a dummy module with imports, a solution that would be independent of the simulator being used.