Extract class for string utils
SVUnit needs to be able to do string manipulation, notably for parsing filter expressions. It needs functions that are not available in the SV language.

We want SVUnit to be easy to use and part of this is that it doesn't require external dependencies. Even if we were to use any kind of external dependencies, this would prevent those libraries from using SVUnit for testing, because of circular dependencies.

While we could import string methods from C via DPI, this puts us at the mercy of the simulators. The mechanics of DPI imports are pretty simulator specific, plus not all simulators might support the same features.

The most robust approach is to keep the required code in the library.