Fix compile error due to reference to variable defined later in package
`test` is defined before the global variable, which QuestaSim doesn't like. Converting this variable to a function doesn't help either. Wrapping it in a class allows for breaking the loop using a forward typedef.

It would have also been possible to declare the package variable before the include of `test.svh`, but this would have looked awkward.