Fix Verilator lint warning related to usage of C++ keyword as identifier
Verilator exits by default when there are lint errors. Users might rely on this, so we don't want to use the ` -Wno-lint` option.

We've used `c` to refer to single character strings in other parts of the code base before and we also use `s` to refer to `strings`, so this change is also worth making regardless of Verilator.