Add Verilator run step
By using `tee` to print to both console and file, we are forced do squash stdout and stderr together. This is not so nice from a user point of view, because they might still expect messages to show up on the output stream where a native Verilator run would print them. Later we might implement a better way of capturing the output to a log file.