Add workaround for Verilator compile errors related to  `disable fork`
Not disabling the fork will not cause any problems functionally, because the processes for the "real" test and for the watchdog that waits for error will just hang forever when the other finishes. One of the following two things could happen:

- the test runs without an error, then the watchdog blocks; if the next test issues an error, this will unblock the current test's watchdog as well
- the tet runs info an error, then the watchdog unblocks; the test can never unblock

The blocked processes will take up memory. For small test runs, this shouldn't be a problem.

The `#0` is also problematic. It seems that it is required for something related to calling the `FAIL_*` macros from outside of an `SVTEST`.

We should come back and fix it once Verilator supports `disable fork`: https://github.com/verilator/verilator/issues/4125
We'll track the issue here: https://github.com/svunit/svunit/issues/236