Move stringification macro to defines file marked as internal
Since this is not part of the public API, we want to do a bit more to "hide" it from the user. If `svunit_pkg.sv` is compiled in an own compilation unit, user code won't even get the stringify macro, because it will only include `svunit_defines.svh`. In case everything is compiled as a single compilation unit, at least we have the `__` and the `svunit` prefixes to avoid name clashes.