Define exit code for internal execution error
We've used the exit codes that pytest defines as inspiration when defining the exit code for command-line usage error. We might as well continue the convention.

Internal execution error means that the command started by the script doesn't execute successfully. This could be because of a compile error, but it could also be because the tool exits with a non-zero code when an `assert` fails. We can't distinguish between the two.