Fix fails in tests that try to start simulator without ensuring that it's present
These tests were trying to start the simulator, but they don't really need to, as they test behavior related to the scripts. The executables weren't found, causing the command to fail, but in the past this was never checked.

The simulator we use is irrelevant. It's easier to mock out Xcelium as it calls `xrun`, whereas QuestSim would need us to mock out `vlib`, `vlog` and `vsim`.