Fix command line option taken by `xvlog` for specifying include directories
All other tools accept `+incdir+`, the command line option of the original Verilog-XL tool, which is more or less a de facto standard nowadays. Vivado Simulator has to be a little more special.

We're patching the output of `buildSVUnit` so that we don't have to add another command line option to that it. That script already does a bit much (generates framework files, collects tests, generates the argument file, etc.) so we don't want to burden it with handling tool differences as well.