Encapsulate code in an`SVTEST/SVTEST_END` block in an `svunit_test`
Since the actual test code is now in a separate `run()` task, we can fuse the test execution part of `SVTEST` and `SVTEST_END` to `SVUNIT_TESTS_BEGIN`.  This code could also be moved to `svunit_testcase` itself, as it does not rely on anything from the module, aside from `svunit_ut`.

We have to ensure that the `svunit_ut` object was created before we can add tests to it. This is normally after `build()` was called.